requirements.txt
attrs==24.2.0
azure-functions==1.13.2
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
pytest-snapshot==0.9.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
urllib3==2.2.3
